Create input and add fonts,layout,main, reset, variable , rename image folder,
