+++ src/assets/styles/fonts.scss
@@ -1,0 +1,50 @@
+// Noto Sarif Normal
+
+@font-face {
+  font-family: 'Noto Serif';
+  font-weight: 400;
+  src: url("../fonts/Noto_Serif/NotoSerif-Regular.ttf") format("ttf"),
+  url("../fonts/Noto_Serif/NotoSerif-Regular.otf") format("otf"),
+  url("../fonts/Noto_Serif/NotoSerif-Regular.woff") format("woff"),
+  url("../fonts/Noto_Serif/NotoSerif-Regular.woff2") format("woff2"),
+  url("../fonts/Noto_Serif/NotoSerif-Regular.eot") format("eot");
+}
+
+// Noto Sarif Bold
+
+@font-face {
+  font-family: 'Noto Serif';
+  font-weight: 700;
+  src: url("../fonts/Noto_Serif/NotoSerif-Bold.ttf") format("truetype"),
+  url("../fonts/Noto_Serif/NotoSerif-Bold.otf") format("otf"),
+  url("../fonts/Noto_Serif/NotoSerif-Bold.woff") format("woff"),
+  url("../fonts/Noto_Serif/NotoSerif-Bold.woff2") format("woff2"),
+  url("../fonts/Noto_Serif/NotoSerif-Bold.eot") format("eot");
+}
+
+// Roboto Normal
+
+@font-face {
+  font-family: 'Roboto Serif';
+  font-weight: 400;
+  src: url("../fonts/Roboto/Roboto-Regular.ttf") format("ttf"),
+  url("../fonts/Roboto/Roboto-Regular.eot") format("eot"),
+  url("../fonts/Roboto/Roboto-Regular.otf") format("otf"),
+  url("../fonts/Roboto/Roboto-Regular.woff") format("woff"),
+  url("../fonts/Roboto/Roboto-Regular.woff2") format("woff2");
+}
+
+// Roboto Bold
+
+@font-face {
+  font-family: 'Roboto Serif';
+  font-weight: 700;
+  src: url("../fonts/Roboto/Roboto-Bold.ttf") format("ttf"),
+  url("../fonts/Roboto/Roboto-Bold.eot") format("eof"),
+  url("../fonts/Roboto/Roboto-Bold.otf") format("otf"),
+  url("../fonts/Roboto/Roboto-Bold.woff") format("woff"),
+  url("../fonts/Roboto/Roboto-Bold.woff2") format("woff2");
+}
+
+
+
