+++ src/assets/styles/layout.scss
@@ -1,0 +1,45 @@
+@import "variable";
+
+@mixin breakpoint($point) {
+  @if $point==desktop {
+    @media (min-width: 901px) {
+      @content;
+    }
+  } @else if $point==tablet {
+    @media (min-width: 601px) {
+      @content;
+    }
+  } @else if $point==tabletOnly {
+    @media (min-width: 601px) and (max-width: 900px) {
+      @content;
+    }
+  } @else if $point==mobileOnly {
+    @media (max-width: 600px) {
+      @content;
+    }
+  }
+}
+
+@mixin container-mobile {
+  @include breakpoint(mobileOnly) {
+    margin: 0 $container-mobile-margin;
+  }
+}
+
+@mixin container-tablet {
+  @include breakpoint(tabletOnly) {
+    margin: 0 $container-tablet-margin;
+  }
+}
+
+@mixin container-desktop {
+  @include breakpoint(desktop) {
+    margin: 0 $container-desktop-margin;
+  }
+}
+
+@mixin container {
+  @include container-mobile;
+  @include container-tablet;
+  @include container-desktop;
+}
