+++ src/assets/styles/main.scss
@@ -1,0 +1,2 @@
+@import "reset";
+@import "fonts";
