+++ src/assets/styles/typography.scss
@@ -1,0 +1,71 @@
+@import "layout";
+
+@mixin noto-font {
+  font-family: 'Noto Serif', serif;
+  font-style: normal;
+  font-stretch: normal;
+}
+
+@mixin roboto-font {
+  font-family: 'Roboto Serif', serif;
+  font-style: normal;
+  font-stretch: normal;
+}
+
+@mixin h1 {
+  
+  @include noto-font;
+
+  font-size: 48px;
+  line-height: 65px;
+  font-weight: 700;
+}
+
+@mixin h2 {
+
+  @include noto-font;
+
+  font-size: 28px;
+  line-height: 38px;
+  font-weight: 700;
+}
+
+@mixin h3 {
+
+  @include noto-font;
+
+  font-size: 24px;
+  line-height: 33px;
+}
+
+@mixin h4 {
+
+  @include noto-font;
+
+  font-size: 20px;
+  line-height: 27px;
+  font-weight: 700;
+}
+
+@mixin h5 {
+  @include noto-font;
+
+  font-size: 18px;
+  line-height: 25px;
+  font-weight: 700;
+}
+
+@mixin h6 {
+  @include noto-font;
+
+  font-size: 14px;
+  line-height: 17px;
+}
+
+@mixin p {
+
+  @include roboto-font;
+
+  font-size: 22px;
+  line-height: 27px;
+}
