+++ src/assets/styles/variable.scss
@@ -1,0 +1,13 @@
+// Colors
+$color-light: #fff;
+$primary-gray: #6e6e6e;
+$secondary-gray: #e3e3e3;
+$color-dark: #191919;
+$color-green: #57CA4A;
+
+
+// Margins
+$container-mobile-margin: 9%;
+$container-tablet-margin: 10%;
+$container-desktop-margin: 8%;
+
